Database migration fixes.

+++ backend/database/migrations/2019_04_09_181100_create_blog_articles_table.php
@@ -15,7 +15,7 @@ class CreateBlogArticlesTable extends Migration
     {
         Schema::create('blog_articles', function (Blueprint $table) {
             $table->bigIncrements('id');
-          $table->unsignedInteger("author")
+          $table->unsignedBigInteger("author")
             ->nullable();
           $table->foreign("author")
             ->references("id")
